feat: adds tests for sendReply
